File download issue, delete issue and copy module file issues are fixed.

+++ app/Http/Controllers/DocumentController.php
@@ -43,7 +43,10 @@ class DocumentController extends Controller {
 
     public function destroy($doc_id) {
         $doc_file = document::where('id', $doc_id)->first();
-        unlink(public_path('documents/' . $doc_file->filename));
+        
+        if(file_exists(public_path('documents/' . $doc_file->filename))) {
+            unlink(public_path('documents/' . $doc_file->filename));            
+        }
         $doc = document::destroy($doc_id);
         return response()->json($doc);
     }
